Fixing bug with segfaulting TransferFromImageBitmap(null)
There was an issue with OffscreenCanvas created from transferring
control from a onscreen canvas, and using a BitmapRenderer context
with the TransferFromImageBitmap(null).

According to the standard TransferFromImageBitmap(null) has to reset the
internal bitmap and create a black transparent one.
https://html.spec.whatwg.org/multipage/canvas.html#the-imagebitmaprenderingcontext-interface

This CL also adds new tests, and fixes some issues with naming in the
tests.

This CL also moves a method that should have been protected and not
public in the first place.

Bug: 1188892
Change-Id: I79f5487c99618fa0bbaf8c436b710766f82ce657